--- src/main/java/wooteco/subway/AdviceController.java
@@ -1,15 +1,0 @@
-package wooteco.subway;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestController;
-
-@ControllerAdvice(annotations = RestController.class)
-public class AdviceController {
-
-    @ExceptionHandler({IllegalArgumentException.class})
-    public ResponseEntity<String> handle(Exception e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-}
