+++ src/main/java/wooteco/subway/domain/Station.java
@@ -1,0 +1,25 @@
+package wooteco.subway.domain;
+
+public class Station {
+
+    private Long id;
+    private String name;
+
+    public Station(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
+
