+++ src/main/java/wooteco/subway/dto/StationRequest.java
@@ -1,0 +1,17 @@
+package wooteco.subway.dto;
+
+public class StationRequest {
+
+    private String name;
+
+    public StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
