+++ src/main/java/wooteco/subway/entity/LineEntity.java
@@ -1,0 +1,30 @@
+package wooteco.subway.entity;
+
+public class LineEntity {
+
+    private Long id;
+    private String name;
+    private String color;
+
+    public LineEntity(Long id, String name, String color) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+    }
+
+    public LineEntity(String name, String color) {
+        this(null, name, color);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
