+++ src/main/java/wooteco/subway/entity/SectionEntity.java
@@ -1,0 +1,46 @@
+package wooteco.subway.entity;
+
+public class SectionEntity {
+    private final Long id;
+    private final Long lineId;
+    private final Long upStationId;
+    private final Long downStationId;
+    private final int distance;
+
+    public SectionEntity(Long id, Long lineId, Long upStationId, Long downStationId, int distance) {
+        this.id = id;
+        this.lineId = lineId;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public SectionEntity(Long lineId, Long upStationId, Long downStationId, int distance) {
+        this(null, lineId, upStationId, downStationId, distance);
+    }
+
+    public SectionEntity(Long id, SectionEntity sectionEntity) {
+        this(id, sectionEntity.getLineId(), sectionEntity.getUpStationId(),
+            sectionEntity.getDownStationId(), sectionEntity.getDistance());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getLineId() {
+        return lineId;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
