+++ src/main/java/wooteco/subway/entity/StationEntity.java
@@ -1,0 +1,24 @@
+package wooteco.subway.entity;
+
+public class StationEntity {
+
+    private Long id;
+    private String name;
+
+    public StationEntity(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public StationEntity(String name) {
+        this(null, name);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
