--- src/main/java/wooteco/subway/station/Station.java
@@ -1,25 +1,0 @@
-package wooteco.subway.station;
-
-public class Station {
-
-    private Long id;
-    private String name;
-
-    public Station(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Station(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
-
