--- src/main/java/wooteco/subway/station/StationEntity.java
@@ -1,24 +1,0 @@
-package wooteco.subway.station;
-
-public class StationEntity {
-
-    private Long id;
-    private String name;
-
-    public StationEntity(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public StationEntity(String name) {
-        this(null, name);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
