--- src/main/java/wooteco/subway/station/StationResponse.java
@@ -1,27 +1,0 @@
-package wooteco.subway.station;
-
-public class StationResponse {
-
-    private Long id;
-    private String name;
-
-    public StationResponse() {
-    }
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public StationResponse(StationEntity stationEntity) {
-        this(stationEntity.getId(), stationEntity.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
